requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.0
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
psutil==5.6.3
pylint==2.3.1
python-engineio==3.9.3
python-socketio==4.3.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
